Rename param_flag_list to signature
